modalbed/requirements.txt
# torch
# torchvision
# torchaudio
# backpack-for-pytorch
numpy
wilds
tqdm==4.66.4
imageio==2.9.0
gdown
parameterized==0.9.0
Pillow==10.3.0
timm==0.9.16
fire
pytorchvideo @ git+https://github.com/facebookresearch/pytorchvideo.git@28fe037d212663c6a24f373b94cc5d478c8c1a1d
easydict==1.9
ftfy==6.0.1
regex==2022.10.31
pyyaml_env_tag==0.1
h5py
ipdb
nvitop
wandb
open3d==0.16.0
termcolor==1.1.0
lmdb==1.3.0
einops
ninja
fairscale
fire
sentencepiece
transformers==4.33.0
timm
KNN-CUDA @ https://github.com/unlimblue/KNN_CUDA/releases/download/0.2/KNN_CUDA-0.2-py3-none-any.whl#sha256=809a43da2eed950c65ba15e355e8cdbf975b6a948dced1f02414f26aba586cd0
tensorboard
peft
opencv-python
decord
